MaxNumber at first Position using float

+++ TestCase.cs
@@ -8,7 +8,7 @@ namespace FindMaximumProblem
 {
     public class TestCase
     {
-        public static int MaxNumber(int firstNumber, int secondNumber, int thirdNumber)
+        public static float MaxNumber(float firstNumber, float secondNumber, float thirdNumber)
         {
             if (firstNumber.CompareTo(secondNumber) > 0 && firstNumber.CompareTo(thirdNumber) > 0)
             {
